Added homepages model and moved top ten methods there
