+++ app/assets/stylesheets/application.css
@@ -53,7 +53,6 @@ footer {
   background-color: #D5CAD6;
   grid-column: span 2;
   height: 50px;
-  /* position: fixed; */
   bottom: 0;
   width: 100%;
 }
