+++ app/controllers/homepages_controller.rb
@@ -7,20 +7,4 @@ class HomepagesController < ApplicationController
     top_ten_books
     top_ten_albums
   end
-
-  def top_ten_movies
-    @movies = Work.where(category: "movie")
-    @top_ten_movies = @movies.sample(10)
-  end
-
-  def top_ten_books
-    @books = Work.where(category: "book")
-    @top_ten_books = @books.sample(10)
-  end
-
-  def top_ten_albums  
-    @albums = Work.where(category: "album")
-    @top_ten_albums = @albums.sample(10)
-  end
-
 end
