+++ db/migrate/20190423212243_create_homepages.rb
@@ -1,0 +1,8 @@
+class CreateHomepages < ActiveRecord::Migration[5.2]
+  def change
+    create_table :homepages do |t|
+
+      t.timestamps
+    end
+  end
+end
